resolve PR comment
